Roll back refactor - everything back to run_capsule.py
